Use log_in_as separately between SystemSpec and RequestSpec
