+++ spec/support/request/helper.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module RequestSpecHelper
+  def log_in_as(user)
+    post login_path, params: { session: { email: user.email, password: "password" } }
+  end
+end
