+++ spec/support/system/helper.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module SystemSpecHelper
+  def log_in_as(user)
+    visit login_path
+    fill_in "Email",    with: user.email
+    fill_in "Password", with: user.password
+    click_button "Log in"
+  end
+end
